权限校验提前至Controller

+++ course/src/main/java/xo/fredtan/lottolearn/course/service/ChapterServiceImpl.java
@@ -8,16 +8,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
 import xo.fredtan.lottolearn.api.course.service.ChapterService;
-import xo.fredtan.lottolearn.common.exception.ApiExceptionCast;
 import xo.fredtan.lottolearn.common.model.response.BasicResponseData;
-import xo.fredtan.lottolearn.common.model.response.CommonCode;
 import xo.fredtan.lottolearn.common.model.response.QueryResponseData;
 import xo.fredtan.lottolearn.common.model.response.QueryResult;
 import xo.fredtan.lottolearn.course.dao.ChapterRepository;
-import xo.fredtan.lottolearn.course.utils.WithUserValidationUtils;
 import xo.fredtan.lottolearn.domain.course.Chapter;
 import xo.fredtan.lottolearn.domain.course.request.ModifyChapterRequest;
-import xo.fredtan.lottolearn.domain.course.response.CourseCode;
 
 import java.util.Date;
 
@@ -26,14 +22,8 @@ import java.util.Date;
 public class ChapterServiceImpl implements ChapterService {
     private final ChapterRepository chapterRepository;
 
-    private final WithUserValidationUtils withUserValidationUtils;
-
     @Override
     public QueryResponseData<Chapter> findChaptersByCourseId(Integer page, Integer size, String courseId) {
-        if (withUserValidationUtils.notParticipate(courseId)) {
-            ApiExceptionCast.cast(CourseCode.NOT_JOIN_COURSE);
-        }
-
         PageRequest pageRequest = PageRequest.of(page, size);
         Page<Chapter> chapters = chapterRepository.findByCourseIdOrderByPubDateDesc(pageRequest, courseId);
 
@@ -44,10 +34,6 @@ public class ChapterServiceImpl implements ChapterService {
     @Override
     @Transactional
     public BasicResponseData addChapter(String courseId, ModifyChapterRequest modifyChapterRequest) {
-        if (withUserValidationUtils.notCourseOwner(courseId)) {
-            ApiExceptionCast.cast(CommonCode.FORBIDDEN);
-        }
-
         Chapter chapter = new Chapter();
         BeanUtils.copyProperties(modifyChapterRequest, chapter);
         chapter.setId(null);
@@ -64,10 +50,6 @@ public class ChapterServiceImpl implements ChapterService {
     public BasicResponseData updateChapter(String courseId,
                                            String chapterId,
                                            ModifyChapterRequest modifyChapterRequest) {
-        if (withUserValidationUtils.notCourseOwner(courseId)) {
-            ApiExceptionCast.forbidden();
-        }
-
         chapterRepository.findById(chapterId).ifPresent(chapter -> {
             // 确保课程ID和发布时间一致
             modifyChapterRequest.setCourseId(chapter.getCourseId());
@@ -82,13 +64,11 @@ public class ChapterServiceImpl implements ChapterService {
 
     @Override
     @Transactional
-    public BasicResponseData deleteChapter(String chapterId) {
+    public BasicResponseData deleteChapter(String courseId, String chapterId) {
         chapterRepository.findById(chapterId).ifPresent(chapter -> {
-            String courseId = chapter.getCourseId();
-            if (withUserValidationUtils.notCourseOwner(courseId)) {
-                ApiExceptionCast.forbidden();
+            if (chapter.getCourseId().equals(courseId)) {
+                chapterRepository.delete(chapter);
             }
-            chapterRepository.delete(chapter);
         });
 
         return BasicResponseData.ok();
